refactor: extract diagram rendering to dedicated class
